fastapi-full-stack-web-development-in-2024/requirements.txt
fastapi
uvicorn
pydantic[email]
httpx

SQLAlchemy
psycopg2

alembic

python-dotenv

passlib

python-slugify

python-jose[cryptography]
python-multipart

jinja2
